requirements.txt
aiohttp==3.8.4
aioredis==1.3.1
aiosignal==1.3.1
annotated-types==0.6.0
ansi2html==1.8.0
appdirs==1.4.4
arabic-reshaper==2.1.4
asgiref==3.5.2
asn1crypto==1.5.1
asttokens==2.2.1
async-generator==1.10
async-timeout==4.0.2
attrs==23.1.0
audioread==3.0.1
autobahn==21.3.1
Automat==20.2.0
backports.zoneinfo==0.2.1
beautifulsoup4==4.12.2
bidict==0.22.1
blinker==1.6.2
bokeh==2.4.2
branca==0.6.0
bs4==0.0.1
cachetools==5.3.2
certifi==2023.7.22
cffi==1.15.1
channels==3.0.3
channels-redis==3.2.0
chardet==4.0.0
charset-normalizer==3.2.0
click==8.1.7
clicksend-client==5.0.72
colorama==0.4.6
constantly==15.1.0
cryptography==38.0.3
cssselect==1.2.0
cssselect2==0.7.0
daphne==3.0.1
dash==2.11.1
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-table==5.0.0
decorator==5.1.1
defusedxml==0.7.1
distlib==0.3.6
Django==4.0
django-allauth==0.54.0
django-background-tasks==1.2.5
django-compat==1.0.15
django-cors-headers==3.6.0
django-crispy-forms==1.11.0
django-crontab==0.7.1
django-debug-toolbar==3.2
django-environ==0.8.1
django-extensions==3.1.3
django-htmx==1.6.0
django-oauth-toolkit==1.3.3
django-pg-bulk-update==3.4.2
django-tinymce==3.2.0
django-utils-six==2.0
django-widget-tweaks==1.4.9
djangorestframework==3.12.2
docopt==0.6.2
einops==0.4.1
encoding-tools==0.0.3
et-xmlfile==1.1.0
exceptiongroup==1.1.3
executing==1.2.0
fake-useragent==1.2.1
feedparser==6.0.10
ffmpeg==1.4
filelock==3.12.0
Flask==2.3.3
Flask-SocketIO==5.3.6
folium==0.14.0
frozendict==2.3.8
frozenlist==1.3.3
fsspec==2023.12.0
future==0.18.2
geographiclib==2.0
geopy==2.3.0
google-api-core==2.14.0
google-api-python-client==2.108.0
google-auth==2.23.4
google-auth-httplib2==0.1.1
google-cloud-bigquery==3.13.0
google-cloud-core==2.3.3
google-crc32c==1.5.0
google-resumable-media==2.6.0
googleapis-common-protos==1.61.0
grpcio==1.59.3
grpcio-status==1.59.3
h11==0.14.0
hiredis==1.1.0
html5lib==1.1
httplib2==0.22.0
huggingface-hub==0.19.4
hyperlink==21.0.0
idna==3.4
importlib-metadata==6.8.0
importlib-resources==6.0.1
incremental==21.3.0
inflect==7.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
joblib==1.3.2
librosa==0.9.1
llvmlite==0.36.0
lxml==4.9.1
MarkupSafe==2.1.3
mpmath==1.3.0
msgpack==1.0.2
multidict==6.0.4
multitasking==0.0.11
ndg-httpsclient==0.5.1
nest-asyncio==1.5.6
networkx==3.1
numba==0.53.0
numpy==1.20.0
oauthlib==3.2.2
openai==0.27.8
openpyxl==3.1.2
oscrypto==1.3.0
outcome==1.2.0
packaging==23.1
pandas==2.0.3
parse==1.19.1
parso==0.8.3
pbr==5.11.1
Pillow==9.2.0
pipreqs==0.4.13
platformdirs==3.5.1
plotly==5.15.0
pooch==1.8.0
progressbar==2.5
proto-plus==1.22.3
protobuf==4.25.1
psycopg2==2.9.4
psycopg2-binary==2.8.6
pure-eval==0.2.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pydantic==2.5.2
pydantic_core==2.14.5
pyee==8.2.2
Pygments==2.15.1
pyHanko==0.15.1
pyhanko-certvalidator==0.19.6
PyJWT==2.6.0
pyOpenSSL==20.0.1
pyparsing==3.1.1
PyPDF3==1.0.6
pyppeteer==1.0.2
PyQt5==5.15.9
PyQt5-Qt5==5.15.2
PyQt5-sip==12.12.2
PyQtWebEngine==5.15.6
PyQtWebEngine-Qt5==5.15.2
pyquery==2.0.0
PySocks==1.7.1
pytesseract==0.3.10
python-bidi==0.4.2
python-dateutil==2.8.2
python-decouple==3.8
python-dotenv==1.0.0
python-engineio==4.8.0
python-socketio==5.10.0
python3-openid==3.2.0
pytz==2023.3.post1
pytz-deprecation-shim==0.1.0.post0
PyVirtualDisplay==3.0
PyYAML==6.0
qrcode==7.3.1
regex==2023.10.3
reportlab==3.6.12
requests==2.31.0
requests-html==0.10.0
requests-oauthlib==1.3.1
resampy==0.4.2
retrying==1.3.4
rotary-embedding-torch==0.3.6
rsa==4.9
safetensors==0.4.1
schedule==1.0.0
scikit-learn==1.3.2
scipy==1.5.4
selenium==4.12.0
service-identity==18.1.0
sgmllib3k==1.0.0
simple-websocket==1.0.0
simplejson==3.17.2
six==1.16.0
sniffio==1.3.0
social-auth-app-django==5.2.0
social-auth-core==4.4.2
sortedcontainers==2.4.0
soundfile==0.12.1
soupsieve==2.4.1
sqlparse==0.4.3
stack-data==0.6.2
stevedore==5.1.0
svglib==1.4.1
sympy==1.12
tenacity==8.2.2
threadpoolctl==3.2.0
tinycss2==1.2.1
tokenizers==0.12.1
torch==2.1.1
torchaudio==2.1.1
tornado==6.3.2
tortoise-tts==3.0.0
tqdm==4.65.0
traitlets==5.9.0
transformers==4.31.0
trio==0.22.2
trio-websocket==0.10.4
tweepy==4.14.0
twilio==7.15.0
Twisted==21.2.0
twisted-iocpsupport==1.0.3
txaio==21.2.1
typing_extensions==4.6.3
tzdata==2023.3
tzlocal==4.2
Unidecode==1.3.7
uritemplate==4.1.1
uritools==4.0.0
urllib3==2.0.4
virtualenv==20.23.0
virtualenv-clone==0.5.7
virtualenvwrapper==4.8.4
virtualenvwrapper-win==1.2.7
w3lib==2.1.2
wcwidth==0.2.6
webdriver-manager==4.0.0
webencodings==0.5.1
websockets==10.4
Werkzeug==2.3.7
wsproto==1.2.0
xhtml2pdf==0.2.8
xvfbwrapper==0.2.9
yahoo-fin==0.8.9.1
yahoo-finance==1.4.0
yarg==0.1.9
yarl==1.9.2
yfinance==0.2.28
zipp==3.16.2
zope.interface==5.3.0
